refactor(ui): flatten website components/marketing/ into components/
The "marketing" subfolder was tautological — the website itself is the
marketing site. All 9 components + barrel + tests move up one level,
every consumer imports through `@/components`, and split imports
(barrel + direct path) are merged into a single barrel import.

The barrel export test added no value beyond what `tsc -b` already
catches and was removed.

Fixes: shellhub-io/team#185
