+++ ui/apps/website/src/components/__tests__/ActionButton.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import { ActionButton, ActionButtonGroup } from "@/components/marketing";
+import { ActionButton, ActionButtonGroup } from "@/components";
 
 function renderButton(
   props?: Partial<React.ComponentProps<typeof ActionButton>>,
