+++ ui/apps/website/src/components/__tests__/InfoCard.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
-import { InfoCard, type InfoCardProps } from "@/components/marketing/InfoCard";
+import { InfoCard, type InfoCardProps } from "@/components";
 import { C } from "@shellhub/design-system/constants";
 
 const color = C.primary;
