+++ ui/apps/website/src/components/__tests__/SectionHeader.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
-import { SectionHeader } from "@/components/marketing/SectionHeader";
+import { SectionHeader } from "@/components";
 
 vi.mock("@shellhub/design-system/components", () => ({
   Reveal: ({
