--- ui/apps/website/src/components/marketing/index.ts
@@ -1,20 +1,0 @@
-export { ActionButton, ActionButtonGroup } from "./ActionButton";
-export type {
-  ActionButtonProps,
-  ActionButtonGroupProps,
-  CTAAction,
-} from "./ActionButton";
-export { Section } from "./Section";
-export type { SectionOwnProps, SectionProps } from "./Section";
-export { SectionHeader } from "./SectionHeader";
-export type { SectionHeaderProps } from "./SectionHeader";
-export { FeatureListItem } from "./FeatureListItem";
-export type { FeatureListItemProps } from "./FeatureListItem";
-export { CommandBlock } from "./CommandBlock";
-export type { CommandBlockProps } from "./CommandBlock";
-export { CTABanner } from "./CTABanner";
-export type { CTABannerProps } from "./CTABanner";
-export { InfoCard } from "./InfoCard";
-export type { InfoCardProps } from "./InfoCard";
-export { HighlightCard } from "./HighlightCard";
-export type { HighlightCardColor, HighlightCardProps } from "./HighlightCard";
