+++ ui/apps/website/src/pages/enterprise/AdminPanel.tsx
@@ -1,6 +1,6 @@
 import { CheckIcon } from "@heroicons/react/24/outline";
 import { WindowChrome } from "@shellhub/design-system/primitives";
-import { Section, SectionHeader } from "@/components/marketing";
+import { Section, SectionHeader } from "@/components";
 import { Reveal, ShimmerCard } from "@shellhub/design-system/components";
 
 const capabilities = [
