+++ ui/apps/website/src/pages/enterprise/DeploymentOptions.tsx
@@ -1,7 +1,6 @@
 import { CloudIcon, ShoppingCartIcon } from "@heroicons/react/24/outline";
 import { Badge, Card, IconBadge } from "@shellhub/design-system/primitives";
-import { HighlightCard, Section, SectionHeader } from "@/components/marketing";
-import { FeatureListItem } from "@/components/marketing/FeatureListItem";
+import { FeatureListItem, HighlightCard, Section, SectionHeader } from "@/components";
 import { Reveal, ShimmerCard } from "@shellhub/design-system/components";
 
 export function DeploymentOptions() {
