+++ ui/apps/website/src/pages/enterprise/SecurityFeatures.tsx
@@ -6,7 +6,7 @@ import {
   ShieldCheckIcon,
 } from "@heroicons/react/24/outline";
 import { PlayIcon } from "@heroicons/react/24/solid";
-import { Section, SectionHeader, InfoCard } from "@/components/marketing";
+import { Section, SectionHeader, InfoCard } from "@/components";
 import { C } from "@shellhub/design-system/constants";
 
 const features = [
