+++ ui/apps/website/src/pages/integrations/index.tsx
@@ -25,7 +25,6 @@ import {
   Reveal,
   ShimmerCard,
 } from "@shellhub/design-system/components";
-import { SiteLayout } from "@/components/SiteLayout";
 import {
   ActionButtonGroup,
   CTABanner,
@@ -33,8 +32,9 @@ import {
   InfoCard,
   Section,
   SectionHeader,
+  SiteLayout,
   type CTAAction,
-} from "@/components/marketing";
+} from "@/components";
 import { docsUrl } from "@/links";
 import { C } from "@shellhub/design-system/constants";
