+++ ui/apps/website/src/pages/landing/Architecture.tsx
@@ -1,5 +1,4 @@
-import { Section, SectionHeader } from "@/components/marketing";
-import { ArrowMarker } from "@/components/marketing/ArrowMarker";
+import { ArrowMarker, Section, SectionHeader } from "@/components";
 import { Reveal } from "@shellhub/design-system/components";
 import { C, FONT_SANS, FONT_MONO } from "@shellhub/design-system/constants";
