+++ ui/apps/website/src/pages/landing/OpenSource.tsx
@@ -1,6 +1,6 @@
 import { StarIcon } from "@heroicons/react/24/solid";
 import { GithubIcon } from "@shellhub/design-system/primitives";
-import { ActionButton, Section, SectionHeader } from "@/components/marketing";
+import { ActionButton, Section, SectionHeader } from "@/components";
 import { Reveal } from "@shellhub/design-system/components";
 import { C } from "@shellhub/design-system/constants";
 import { githubUrl, signupUrl } from "@/links";
