+++ ui/apps/website/src/pages/landing/QuickStart.tsx
@@ -1,5 +1,5 @@
 import { ArrowRightIcon } from "@heroicons/react/24/outline";
-import { CommandBlock, Section, SectionHeader } from "@/components/marketing";
+import { CommandBlock, Section, SectionHeader } from "@/components";
 import { Reveal } from "@shellhub/design-system/components";
 import { docsUrl } from "@/links";
