+++ ui/apps/website/src/pages/use-cases/DevopsCiCd.tsx
@@ -15,15 +15,15 @@ import {
   Reveal,
   ShimmerCard,
 } from "@shellhub/design-system/components";
-import { SiteLayout } from "@/components/SiteLayout";
 import {
   ActionButtonGroup,
   CTABanner,
   InfoCard,
   Section,
   SectionHeader,
+  SiteLayout,
   type CTAAction,
-} from "@/components/marketing";
+} from "@/components";
 import { docsUrl } from "@/links";
 import { C } from "@shellhub/design-system/constants";
