+++ ui/apps/website/src/pages/use-cases/RemoteSupport.tsx
@@ -16,15 +16,15 @@ import {
   Reveal,
   ShimmerCard,
 } from "@shellhub/design-system/components";
-import { SiteLayout } from "@/components/SiteLayout";
 import {
   ActionButtonGroup,
   CTABanner,
   InfoCard,
   Section,
   SectionHeader,
+  SiteLayout,
   type CTAAction,
-} from "@/components/marketing";
+} from "@/components";
 import { C } from "@shellhub/design-system/constants";
 
 const primaryAction: CTAAction = {
